add stake_test file
